Re-organize project to match article PRs.
