+++ app/src/main/java/com/robotandpencils/app/ui/common/BaseFragmentModule.java
@@ -1,4 +1,4 @@
-package com.robotandpencils.app.ui.common.view;
+package com.robotandpencils.app.ui.common;
 
 import android.app.Fragment;
 import android.app.FragmentManager;
