+++ app/src/main/res/layout/main_activity.xml
@@ -14,10 +14,18 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   -->
-<resources>
-    <string name="app_name">Android Dagger ButterKnife MVP</string>
 
-    <string name="example_1">Ex1: Activity with 1 Fragment</string>
-    <string name="example_2">Ex2: Activity with 2 Fragments</string>
-    <string name="example_3">Ex3: Activity with 1 Fragment with 1 child Fragment</string>
-</resources>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+             android:layout_width="match_parent"
+             android:layout_height="match_parent">
+
+    <!-- We don't need this inner FrameLayout. We have it here just for consistency with the rest
+        of the sample app -->
+    <!-- We can also embed the fragment in xml using the fragment tag. However, we create the
+        fragment in Java for this example. -->
+    <FrameLayout
+        android:id="@+id/fragment_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
+
+</FrameLayout>
