+++ app/src/main/res/layout/main_fragment.xml
@@ -1,0 +1,45 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright 2017 Vandolf Estrellado
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~        http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <Button
+            android:id="@+id/example_1"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/example_1"/>
+
+        <Button
+            android:id="@+id/example_2"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/example_2" />
+
+        <Button
+            android:id="@+id/example_3"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="@string/example_3" />
+
+    </LinearLayout>
+</ScrollView>
